pyproject.toml
[project]
name = "researcher-agent"
version = "0.3.0"
dependencies = [
    "crewai[tools]>=0.100.0",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "langchain-community",
    "httpx",
    "Pillow",
    "diffusers",
    "transformers",
    "accelerate",
    "safetensors",
    "aiosqlite",
    "openai-whisper",
    "edge-tts",
    "langdetect",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]